+++ src/lib/classes/appContent.js
@@ -1,0 +1,27 @@
+/**
+  The AppContent object.
+  @class F2.AppManifest.AppContent
+*/
+module.exports = {
+  /**
+    Arbitrary data to be passed along with the app.
+    @property data
+    @type object
+    @optional
+  */
+  data: {},
+  /**
+    The string of HTML representing the app.
+    @property html
+    @type string
+    @required
+  */
+  html: '',
+  /**
+    A status message.
+    @property status
+    @type string
+    @optional
+  */
+  status: ''
+};
