+++ src/lib/classes/index.js
@@ -1,0 +1,13 @@
+var app = require('./app');
+var appConfig = require('./appConfig');
+var appContent = require('./appContent');
+var appManifest = require('./appManifest');
+var containerConfig = require('./containerConfig');
+
+module.exports = {
+  App: app,
+  AppConfig: appConfig,
+  AppContent: appContent,
+  AppManifest: appManifest,
+  ContainerConfig: containerConfig
+};
